benchmark_requirements.txt
Pillow==3.2.0
tabulate==0.7.5
